+++ paddle/phi/kernels/impl/matmul_kernel_impl.h
@@ -1111,8 +1111,8 @@ DispatchMatmulFP8Kernel(const Context& ctx,
   phi::DenseTensor workspace;
   workspace.Resize({30 * 1024 * 1024});
   ctx.template Alloc<int8_t>(&workspace);
-  ctx.template Alloc<phi::dtype::bfloat16>(out);
-  CublasLtMatmulFP8<phi::dtype::float8_e4m3>(ctx, x, y, workspace, out);
+  ctx.template Alloc<phi::dtype::float16>(out);
+  CublasLtMatmulFP8<phi::dtype::float16>(ctx, x, y, workspace, out);
 }
 #endif
